backend/requirements.txt
cryptography>=42
dj-database-url>=2.1.0
django>=5.0
django-allauth>=0.61.0,<0.62.0
django-cors-headers==4.4.0
django-ninja>=1.1.0
django-ninja-extra>=0.19.1
djangorestframework>=3.15.0,<3.16.0
djangorestframework-simplejwt>=5.3.1,<6.0.0
email-validator>=2.0.0
google-auth>=2.32.0,<3.0
google-generativeai
gunicorn>=22.0.0,<23.0.0
opencv-python-headless
Pillow>=10.4.0,<11.0.0
psycopg[binary]>=3.2
PyJWT~=2.9
pytest
pytest-django
python-dotenv>=1.0.1
pytesseract
pyzbar
requests>=2.32.0,<3.0
uvicorn[standard]>=0.30.0,<0.31.0
whitenoise==6.7.0
pyzbar
